リファクタリング: auth.ts / token.ts の大規模ルートファイル分割 (#113)
auth.ts (~970行) と token.ts (~677行) をルートハンドラー単位で分割:

routes/auth/
  ├── index.ts         (ルート定義・ミドルウェア登録)
  ├── authorize.ts     (OAuth 2.0 Authorization)
  ├── login.ts         (BFFリダイレクト + プロバイダー認可)
  ├── callback.ts      (OAuthコールバック)
  ├── exchange.ts      (ワンタイムコード交換)
  ├── refresh.ts       (トークンリフレッシュ)
  ├── link-intent.ts   (SNSプロバイダー連携トークン)
  └── logout.ts        (ログアウト)

routes/token/
  ├── index.ts                    (ルート定義・grant_type振り分け)
  ├── authorization-code-grant.ts (認可コードグラント)
  ├── refresh-token-grant.ts      (リフレッシュトークングラント)
  ├── introspect.ts               (RFC 7662 イントロスペクション)
  ├── revoke.ts                   (RFC 7009 トークン失効)
  └── utils.ts                    (共通型・ヘルパー関数)
