--- workers/id/src/routes/token.ts
@@ -1,677 +1,0 @@
-import { Hono } from "hono";
-import type { HonoRequest } from "hono";
-import {
-  findRefreshTokenByHash,
-  findUserById,
-  revokeRefreshToken,
-  sha256,
-  generatePairwiseSub,
-  verifyAccessToken,
-  createLogger,
-  findAndConsumeAuthCode,
-  findServiceByClientId,
-  generateCodeChallenge,
-  timingSafeEqual,
-  matchRedirectUri,
-  normalizeRedirectUri,
-  addRevokedAccessToken,
-  isAccessTokenRevoked,
-} from "@0g0-id/shared";
-import type { IdpEnv, User, Service, RefreshToken, TokenPayload } from "@0g0-id/shared";
-import {
-  externalApiRateLimitMiddleware,
-  tokenApiClientRateLimitMiddleware,
-  tokenApiRateLimitMiddleware,
-} from "../middleware/rate-limit";
-import { authenticateService } from "../utils/service-auth";
-import { parseAllowedScopes, resolveEffectiveScope } from "../utils/scopes";
-import { handleDeviceCodeGrant } from "./device";
-import { issueTokenPair, buildTokenResponse, issueIdToken } from "../utils/token-pair";
-import { attemptUnrevokeToken } from "../utils/token-recovery";
-import {
-  validateAndRevokeRefreshToken,
-  issueTokenPairWithRecovery,
-} from "../utils/refresh-token-rotation";
-
-const tokenLogger = createLogger("token");
-
-const app = new Hono<{ Bindings: IdpEnv }>();
-
-/**
- * `handleAuthorizationCodeGrant` / `handleRefreshTokenGrant` で共通利用するコンテキスト型。
- * Hono の Context から必要な最小インターフェースのみ抽出。
- */
-export type TokenHandlerContext = {
-  env: IdpEnv;
-  req: { header: (name: string) => string | undefined };
-  json: (data: unknown, status?: number, headers?: Record<string, string>) => Response;
-  header: (name: string, value: string) => void;
-};
-
-/**
- * RFC 7009 / RFC 7662 準拠: リクエストボディのパース。
- * application/x-www-form-urlencoded（RFC標準）と application/json（後方互換）の両方に対応。
- */
-async function parseTokenBody(
-  req: HonoRequest,
-): Promise<{ token?: string; token_type_hint?: string } | null> {
-  const contentType = req.header("Content-Type") ?? "";
-  try {
-    if (contentType.includes("application/x-www-form-urlencoded")) {
-      const body = await req.parseBody();
-      return {
-        token: typeof body["token"] === "string" ? body["token"] : undefined,
-        token_type_hint:
-          typeof body["token_type_hint"] === "string" ? body["token_type_hint"] : undefined,
-      };
-    }
-    return await req.json<{ token?: string; token_type_hint?: string }>();
-  } catch (err) {
-    tokenLogger.error("Failed to parse request body", err);
-    return null;
-  }
-}
-
-/**
- * スコープに基づいてイントロスペクションレスポンスへユーザークレームを付与する。
- * refresh_token / access_token の両ブランチで共通利用。
- */
-function applyUserClaims(claims: Record<string, unknown>, user: User, scopes: string[]): void {
-  if (scopes.includes("profile")) {
-    claims["name"] = user.name;
-    claims["picture"] = user.picture;
-  }
-  if (scopes.includes("email")) {
-    claims["email"] = user.email;
-    claims["email_verified"] = user.email_verified === 1;
-  }
-  if (scopes.includes("phone")) {
-    claims["phone"] = user.phone;
-  }
-  if (scopes.includes("address")) {
-    claims["address"] = user.address;
-  }
-}
-
-/**
- * client_secret_basic 認証（Authorization: Basic）またはパブリッククライアント（none）を処理する。
- * Authorization ヘッダーがある場合は Basic 認証を検証し、クライアントIDの一致も確認する。
- * ヘッダーがない場合はパブリッククライアントとして bodyClientId のみで検証する。
- */
-async function resolveOAuthClient(
-  db: D1Database,
-  authHeader: string | undefined,
-  bodyClientId: string | undefined,
-): Promise<
-  | {
-      ok: true;
-      service: Service;
-      isPublicClient: boolean;
-    }
-  | { ok: false; error: string; status: 400 | 401 | 500 }
-> {
-  if (authHeader?.startsWith("Basic ")) {
-    // Confidential client: client_secret_basic
-    let service: Service | null;
-    try {
-      service = await authenticateService(db, authHeader);
-    } catch (err) {
-      tokenLogger.error("[resolveOAuthClient] Failed to authenticate service", err);
-      return { ok: false, error: "server_error", status: 500 };
-    }
-    if (!service) {
-      return { ok: false, error: "invalid_client", status: 401 };
-    }
-    // bodyのclient_idが指定されている場合、Basicヘッダーのclient_idと一致するか確認
-    if (bodyClientId && bodyClientId !== service.client_id) {
-      return { ok: false, error: "invalid_client", status: 401 };
-    }
-    return { ok: true, service, isPublicClient: false };
-  }
-
-  // Public client: client_id のみで識別（client_secret なし）
-  if (!bodyClientId) {
-    return { ok: false, error: "invalid_request", status: 400 };
-  }
-  let service: Service | null;
-  try {
-    service = await findServiceByClientId(db, bodyClientId);
-  } catch (err) {
-    tokenLogger.error("[resolveOAuthClient] Failed to find service by client_id", err);
-    return { ok: false, error: "server_error", status: 500 };
-  }
-  if (!service) {
-    return { ok: false, error: "invalid_client", status: 401 };
-  }
-  // confidentialクライアント（client_secret_hash設定済み）はBasic認証必須
-  if (service.client_secret_hash) {
-    return { ok: false, error: "invalid_client", status: 401 };
-  }
-  return { ok: true, service, isPublicClient: true };
-}
-
-// POST /api/token — 標準 OAuth 2.0 トークンエンドポイント (RFC 6749)
-// MCPクライアント等のネイティブアプリが直接HTTPリクエストで利用する
-app.post("/", tokenApiRateLimitMiddleware, tokenApiClientRateLimitMiddleware, async (c) => {
-  // RFC 6749 §4.1.3: application/x-www-form-urlencoded のみ受け付ける
-  const contentType = c.req.header("Content-Type") ?? "";
-  if (!contentType.includes("application/x-www-form-urlencoded")) {
-    return c.json(
-      {
-        error: "invalid_request",
-        error_description: "Content-Type must be application/x-www-form-urlencoded",
-      },
-      400,
-    );
-  }
-  let params: Record<string, string>;
-  try {
-    const body = await c.req.parseBody();
-    params = {};
-    for (const [key, value] of Object.entries(body)) {
-      if (typeof value === "string") {
-        params[key] = value;
-      }
-    }
-  } catch {
-    return c.json(
-      { error: "invalid_request", error_description: "Failed to parse request body" },
-      400,
-    );
-  }
-
-  const grantType = params["grant_type"];
-
-  if (grantType === "authorization_code") {
-    return handleAuthorizationCodeGrant(c, params);
-  } else if (grantType === "refresh_token") {
-    return handleRefreshTokenGrant(c, params);
-  } else if (grantType === "urn:ietf:params:oauth:grant-type:device_code") {
-    return handleDeviceCodeGrant(c, params);
-  } else {
-    return c.json(
-      { error: "unsupported_grant_type", error_description: "Unsupported grant_type" },
-      400,
-    );
-  }
-});
-
-/**
- * authorization_code グラント処理
- */
-async function handleAuthorizationCodeGrant(
-  c: TokenHandlerContext,
-  params: Record<string, string>,
-): Promise<Response> {
-  const code = params["code"];
-  const redirectUri = params["redirect_uri"];
-  const clientId = params["client_id"];
-  const codeVerifier = params["code_verifier"];
-
-  if (!code) {
-    return c.json({ error: "invalid_request", error_description: "code is required" }, 400);
-  }
-  if (!redirectUri) {
-    return c.json({ error: "invalid_request", error_description: "redirect_uri is required" }, 400);
-  }
-  // クライアント認証
-  const clientResult = await resolveOAuthClient(c.env.DB, c.req.header("Authorization"), clientId);
-  if (!clientResult.ok) {
-    if (clientResult.status === 401 && clientResult.error === "invalid_client") {
-      c.header("WWW-Authenticate", 'Basic realm="0g0-id"');
-    }
-    return c.json({ error: clientResult.error }, clientResult.status);
-  }
-  const { service, isPublicClient } = clientResult;
-
-  // 認可コード検証
-  const codeHash = await sha256(code);
-  const authCode = await findAndConsumeAuthCode(c.env.DB, codeHash);
-  if (!authCode) {
-    return c.json(
-      { error: "invalid_grant", error_description: "Invalid or expired authorization code" },
-      400,
-    );
-  }
-
-  // service_id の一致確認
-  if (authCode.service_id !== service.id) {
-    return c.json(
-      {
-        error: "invalid_grant",
-        error_description: "Authorization code was not issued for this client",
-      },
-      400,
-    );
-  }
-
-  // redirect_uri を正規化してから比較（RFC 6749 §4.1.3）
-  const normalizedRedirectUri = normalizeRedirectUri(redirectUri);
-  if (!normalizedRedirectUri || !matchRedirectUri(authCode.redirect_to, normalizedRedirectUri)) {
-    return c.json({ error: "invalid_grant", error_description: "redirect_uri mismatch" }, 400);
-  }
-
-  // パブリッククライアントはPKCEを必須とする（RFC 7636 §4.4 / OAuth 2.1）
-  if (isPublicClient && !authCode.code_challenge) {
-    return c.json(
-      { error: "invalid_grant", error_description: "PKCE is required for public clients" },
-      400,
-    );
-  }
-
-  // PKCE 検証 (S256)
-  // code_challenge が存在する場合のみ code_verifier を必須とする（RFC 7636 §4.4）
-  // コンフィデンシャルクライアントは PKCE 不要のため、code_challenge なしでも許容
-  if (authCode.code_challenge) {
-    if (!codeVerifier) {
-      return c.json(
-        { error: "invalid_request", error_description: "code_verifier is required" },
-        400,
-      );
-    }
-    const expectedChallenge = await generateCodeChallenge(codeVerifier);
-    if (!timingSafeEqual(expectedChallenge, authCode.code_challenge)) {
-      return c.json({ error: "invalid_grant", error_description: "code_verifier mismatch" }, 400);
-    }
-  }
-
-  try {
-    // ユーザー情報取得
-    const user = await findUserById(c.env.DB, authCode.user_id);
-    if (!user) {
-      return c.json({ error: "invalid_grant", error_description: "User not found" }, 400);
-    }
-    if (user.banned_at !== null) {
-      return c.json(
-        { error: "access_denied", error_description: "Account has been suspended" },
-        403,
-      );
-    }
-
-    // スコープ計算
-    const serviceScope = resolveEffectiveScope(authCode.scope, service.allowed_scopes);
-    if (serviceScope === undefined) {
-      return c.json({ error: "invalid_scope", error_description: "No valid scope" }, 400);
-    }
-
-    // トークン発行
-    const { accessToken, refreshToken } = await issueTokenPair(c.env.DB, c.env, user, {
-      serviceId: service.id,
-      clientId: service.client_id,
-      scope: serviceScope,
-    });
-
-    // OIDC ID トークン発行（openid スコープがある場合）
-    const idToken = await issueIdToken(c.env, user, service, serviceScope, {
-      nonce: authCode.nonce ?? undefined,
-      amr: authCode.provider ? [authCode.provider] : undefined,
-    });
-
-    // レスポンス (RFC 6749 §5.1)
-    return c.json(buildTokenResponse(accessToken, refreshToken, serviceScope, idToken));
-  } catch (e) {
-    tokenLogger.error("handleAuthorizationCodeGrant: unexpected error", e);
-    return c.json(
-      { error: "server_error", error_description: "An unexpected error occurred" },
-      500,
-    );
-  }
-}
-
-/**
- * refresh_token グラント処理
- */
-async function handleRefreshTokenGrant(
-  c: TokenHandlerContext,
-  params: Record<string, string>,
-): Promise<Response> {
-  const refreshTokenRaw = params["refresh_token"];
-  const clientId = params["client_id"];
-
-  if (!refreshTokenRaw) {
-    return c.json(
-      { error: "invalid_request", error_description: "refresh_token is required" },
-      400,
-    );
-  }
-
-  // クライアント認証
-  const clientResult = await resolveOAuthClient(c.env.DB, c.req.header("Authorization"), clientId);
-  if (!clientResult.ok) {
-    if (clientResult.status === 401 && clientResult.error === "invalid_client") {
-      c.header("WWW-Authenticate", 'Basic realm="0g0-id"');
-    }
-    return c.json({ error: clientResult.error }, clientResult.status);
-  }
-  const service = clientResult.service;
-
-  const tokenHash = await sha256(refreshTokenRaw);
-
-  const validationResult = await validateAndRevokeRefreshToken(c.env.DB, tokenHash);
-  if (!validationResult.ok) {
-    if (validationResult.reason === "TOKEN_ROTATED") {
-      return c.json(
-        { error: "invalid_grant", error_description: "Token rotation in progress, please retry" },
-        400,
-      );
-    }
-    if (validationResult.reason === "TOKEN_REUSE") {
-      return c.json({ error: "invalid_grant", error_description: "Token reuse detected" }, 400);
-    }
-    return c.json({ error: "invalid_grant", error_description: "Invalid refresh token" }, 400);
-  }
-  const storedToken = validationResult.storedToken;
-
-  // サービス所有権確認 & 有効期限チェック（D1クエリを1回に統合）
-  const serviceMismatch = storedToken.service_id !== service.id;
-  const isExpired = new Date(storedToken.expires_at) < new Date();
-
-  if (serviceMismatch || isExpired) {
-    // 並行リクエストが reuse_detected を発動した可能性をチェック（レース条件対策）
-    const currentToken = await findRefreshTokenByHash(c.env.DB, tokenHash);
-    if (currentToken && currentToken.revoked_reason === "reuse_detected") {
-      return c.json({ error: "invalid_grant", error_description: "Token reuse detected" }, 400);
-    }
-    // 期限切れトークン: rotation済み状態のまま拒否（unrevokeは不要かつ危険）
-    // revokedを戻すとreuse detectionロジックが誤動作するリスクがある
-    // NOTE: isExpiredをserviceMismatchより先にチェックする
-    //   両方trueの場合にunrevokeが走ると、期限切れトークンが復活してしまうため
-    if (isExpired) {
-      return c.json({ error: "invalid_grant", error_description: "Refresh token expired" }, 400);
-    }
-    if (serviceMismatch) {
-      // 別サービス向けのトークン → 元に戻して拒否
-      await attemptUnrevokeToken(c.env.DB, storedToken.id, "[token] service_id mismatch 後");
-      return c.json(
-        { error: "invalid_grant", error_description: "Token was not issued for this client" },
-        400,
-      );
-    }
-  }
-
-  // ユーザー情報取得
-  const user = await findUserById(c.env.DB, storedToken.user_id);
-  if (!user) {
-    return c.json({ error: "invalid_grant", error_description: "User not found" }, 400);
-  }
-  if (user.banned_at !== null) {
-    return c.json({ error: "access_denied", error_description: "Account has been suspended" }, 403);
-  }
-
-  // スコープ引き継ぎ
-  const refreshScope =
-    storedToken.scope ?? (parseAllowedScopes(service.allowed_scopes).join(" ") || "openid");
-
-  const issueResult = await issueTokenPairWithRecovery(
-    c.env.DB,
-    c.env,
-    user,
-    {
-      serviceId: service.id,
-      clientId: service.client_id,
-      familyId: storedToken.family_id,
-      scope: refreshScope,
-    },
-    storedToken.id,
-    tokenHash,
-    tokenLogger,
-    "handleRefreshTokenGrant",
-  );
-  if (!issueResult.ok) {
-    if (issueResult.reason === "TOKEN_REUSE") {
-      return c.json({ error: "invalid_grant", error_description: "Token reuse detected" }, 400);
-    }
-    return c.json({ error: "server_error", error_description: "Token operation failed" }, 500);
-  }
-  const accessToken = issueResult.accessToken;
-  const newRefreshToken = issueResult.refreshToken;
-
-  // レスポンス (RFC 6749 §5.1)
-  return c.json(buildTokenResponse(accessToken, newRefreshToken, refreshScope));
-}
-
-// イントロスペクション: リフレッシュトークンの検証ヘルパー（RFC 7662）
-// 見つかった場合はレスポンスオブジェクト（active:false含む）を返し、見つからない・失効済みの場合はnullを返す
-async function introspectRefreshToken(
-  db: D1Database,
-  service: Service,
-  tokenHash: string,
-  issuer: string,
-): Promise<Record<string, unknown> | null> {
-  let refreshToken: RefreshToken | null;
-  let user: User | null | undefined;
-  try {
-    refreshToken = await findRefreshTokenByHash(db, tokenHash);
-    if (!refreshToken) return null;
-    if (refreshToken.revoked_at !== null) return { active: false };
-    if (refreshToken.service_id !== service.id) {
-      tokenLogger.warn(`[introspect] service_id mismatch for requesting service.id=${service.id}`);
-      return { active: false };
-    }
-    if (new Date(refreshToken.expires_at) < new Date()) {
-      return { active: false };
-    }
-    user = await findUserById(db, refreshToken.user_id);
-  } catch (err) {
-    tokenLogger.error("Introspect: DB error in introspectRefreshToken", err);
-    // RFC 7662 §2.2: サーバーエラーは呼び出し側で 500 を返せるよう throw する
-    throw err;
-  }
-  if (!user || user.banned_at !== null) {
-    return { active: false };
-  }
-  const scopeStr = refreshToken.scope ?? "";
-  const scopeList = scopeStr.split(" ").filter((s: string) => s !== "openid" && s !== "");
-  const sub = await generatePairwiseSub(service.client_id, refreshToken.user_id);
-  const response: Record<string, unknown> = {
-    active: true,
-    iss: issuer,
-    token_type: "refresh_token",
-    sub,
-    exp: Math.floor(new Date(refreshToken.expires_at).getTime() / 1000),
-    iat: Math.floor(new Date(refreshToken.created_at).getTime() / 1000),
-    scope: scopeStr,
-  };
-  applyUserClaims(response, user, scopeList);
-  return response;
-}
-
-// イントロスペクション: JWTアクセストークンの検証ヘルパー（RFC 7662）
-// 検証成功時はレスポンスオブジェクトを返し、JWT検証失敗時はnullを返す
-async function introspectJwtToken(
-  db: D1Database,
-  service: Service,
-  token: string,
-  env: IdpEnv,
-): Promise<Record<string, unknown> | null> {
-  let payload: TokenPayload;
-  try {
-    payload = await verifyAccessToken(token, env.JWT_PUBLIC_KEY, env.IDP_ORIGIN, env.IDP_ORIGIN);
-  } catch (err) {
-    tokenLogger.warn("Introspect: JWT verification failed", err);
-    return null;
-  }
-  // RFC 7009: jtiがブロックリストに存在する場合は失効済み
-  try {
-    if (payload.jti && (await isAccessTokenRevoked(db, payload.jti))) {
-      return { active: false };
-    }
-    // BFFセッショントークン（cid未設定）は外部サービスからイントロスペクト不可
-    if (!payload.cid || payload.cid !== service.client_id) {
-      return { active: false };
-    }
-    const tokenUser = await findUserById(db, payload.sub);
-    if (!tokenUser || tokenUser.banned_at !== null) {
-      return { active: false };
-    }
-    // JWTに埋め込まれた発行時スコープを優先（リフレッシュトークンブランチとの一貫性）
-    // スコープ未設定の旧トークンは空文字列（RFC 7662 §2.2 - 実際に付与されたスコープのみ返す）
-    const tokenScopeStr = payload.scope ?? "";
-    const tokenScopes = tokenScopeStr.split(" ").filter((s: string) => s !== "openid" && s !== "");
-    const sub = await generatePairwiseSub(service.client_id, payload.sub);
-    const jwtResponse: Record<string, unknown> = {
-      active: true,
-      iss: payload.iss,
-      sub,
-      exp: payload.exp,
-      iat: payload.iat,
-      scope: tokenScopeStr,
-      token_type: "access_token",
-    };
-    applyUserClaims(jwtResponse, tokenUser, tokenScopes);
-    return jwtResponse;
-  } catch (err) {
-    tokenLogger.error("Introspect: DB error in introspectJwtToken", err);
-    // RFC 7662 §2.2: サーバーエラーは呼び出し側で 500 を返せるよう throw する
-    throw err;
-  }
-}
-
-// POST /api/token/introspect — RFC 7662 トークンイントロスペクション
-app.post("/introspect", externalApiRateLimitMiddleware, async (c) => {
-  // Basic認証でサービス認証
-  let service: Service | null;
-  try {
-    service = await authenticateService(c.env.DB, c.req.header("Authorization"));
-  } catch (err) {
-    tokenLogger.error("Introspect: service authentication failed", err);
-    return c.json({ error: "server_error" }, 500);
-  }
-  if (!service) {
-    c.header("WWW-Authenticate", 'Basic realm="0g0-id"');
-    return c.json({ active: false }, 401);
-  }
-  const introspectService = service;
-
-  // トークン取得（RFC 7662: application/x-www-form-urlencoded および application/json に対応）
-  const body = await parseTokenBody(c.req);
-  if (!body) {
-    return c.json({ active: false }, 400);
-  }
-  if (!body.token) {
-    return c.json({ active: false }, 400);
-  }
-
-  const token = body.token;
-  const tokenHash = await sha256(token);
-
-  // token_type_hint に従って検索順を最適化（RFC 7662 §2.1 推奨）
-  // 'access_token' ヒントならJWT→リフレッシュトークンの順、それ以外はリフレッシュトークン→JWTの順
-  let result: Record<string, unknown> | null;
-  try {
-    if (body.token_type_hint === "access_token") {
-      result = await introspectJwtToken(c.env.DB, introspectService, token, c.env);
-      if (result === null) {
-        result = await introspectRefreshToken(
-          c.env.DB,
-          introspectService,
-          tokenHash,
-          c.env.IDP_ORIGIN,
-        );
-      }
-    } else {
-      result = await introspectRefreshToken(
-        c.env.DB,
-        introspectService,
-        tokenHash,
-        c.env.IDP_ORIGIN,
-      );
-      if (result === null) {
-        result = await introspectJwtToken(c.env.DB, introspectService, token, c.env);
-      }
-    }
-  } catch (err) {
-    tokenLogger.error("Introspect: unexpected server error", err);
-    // RFC 7662 §2.2: サーバーエラー時は 500 を返す（{ active: false } との混同を防ぐ）
-    return c.json({ error: "server_error" }, 500);
-  }
-
-  return c.json(result ?? { active: false });
-});
-
-// POST /api/token/revoke — RFC 7009 トークン失効
-app.post("/revoke", externalApiRateLimitMiddleware, async (c) => {
-  // Basic認証でサービス認証
-  let service: Service | null;
-  try {
-    service = await authenticateService(c.env.DB, c.req.header("Authorization"));
-  } catch (err) {
-    tokenLogger.error("Revoke: service authentication failed", err);
-    return c.json({ error: "server_error" }, 500);
-  }
-  if (!service) {
-    c.header("WWW-Authenticate", 'Basic realm="0g0-id"');
-    return c.json({ error: "invalid_client" }, 401);
-  }
-
-  // トークン取得（RFC 7009: application/x-www-form-urlencoded および application/json に対応）
-  const body = await parseTokenBody(c.req);
-  if (!body) {
-    return c.json({ error: "invalid_request" }, 400);
-  }
-
-  if (!body.token) {
-    return c.json({ error: "invalid_request" }, 400);
-  }
-
-  const token = body.token;
-
-  // JWTアクセストークンの失効処理（RFC 7009 §2.1）
-  // JWTは header.payload.signature の3セクション形式（Base64url）で識別する
-  const JWT_PATTERN = /^[A-Za-z0-9\-_]+\.[A-Za-z0-9\-_]+\.[A-Za-z0-9\-_]+$/;
-  if (JWT_PATTERN.test(token)) {
-    let payload: TokenPayload | null = null;
-    try {
-      payload = await verifyAccessToken(
-        token,
-        c.env.JWT_PUBLIC_KEY,
-        c.env.IDP_ORIGIN,
-        c.env.IDP_ORIGIN,
-      );
-    } catch {
-      // JWT検証失敗 → リフレッシュトークン処理へフォールスルー
-    }
-    if (payload !== null) {
-      // 自サービスが発行したトークンかつ有効期限内のものだけブロックリストに追加
-      // cidが未設定の旧BFFセッショントークンはスキップ（introspectと同じ設計）
-      if (
-        payload.jti &&
-        payload.cid &&
-        payload.cid === service.client_id &&
-        payload.exp &&
-        payload.exp > Math.floor(Date.now() / 1000)
-      ) {
-        try {
-          await addRevokedAccessToken(c.env.DB, payload.jti, payload.exp);
-        } catch (err) {
-          tokenLogger.error("Revoke: failed to add revoked access token to blocklist", err);
-          return c.json({ error: "server_error" }, 500);
-        }
-      }
-      return new Response(null, { status: 200 });
-    }
-  }
-
-  // リフレッシュトークンの失効処理
-  const tokenHash = await sha256(token);
-  try {
-    const refreshToken = await findRefreshTokenByHash(c.env.DB, tokenHash);
-
-    // RFC 7009: トークンが存在しない・失効済みでも 200 OK を返す（情報漏洩防止）
-    // 自サービスが発行したトークンのみ失効可能
-    if (
-      refreshToken &&
-      refreshToken.revoked_at === null &&
-      refreshToken.service_id === service.id
-    ) {
-      await revokeRefreshToken(c.env.DB, refreshToken.id, "service_revoke");
-    }
-  } catch (err) {
-    tokenLogger.error("Revoke: failed to process refresh token", err);
-    return c.json({ error: "server_error" }, 500);
-  }
-
-  return new Response(null, { status: 200 });
-});
-
-export default app;
